Add a new extensible storage factory
This new storage factory uses a `FileInstance.storage_backend` attribute to
determine which backend to use. It uses the
`FILES_REST_DEFAULT_STORAGE_BACKEND` config value when the file instance
doesn't already have a storage backend.

It can be extended to customize how the backend is chosen, and how it is
initialised.

The file path is chosen based on the `id` of the file instance, but this can be
customised too.

It's not set as the default storage factory to preserve backwards compatibility.